requirements.txt
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.0
# You must install the MCP SDK for Python, e.g.:
# pip install modelcontextprotocol-mcp-sdk
# If not available, see https://github.com/modelcontextprotocol/mcp-sdk or use your own implementation.
mcp
wave